+++ add_subtitles_to_frames.py
@@ -34,8 +34,8 @@ class AddSubtitlesToFramesNode:
                 }),
                 "center_x": ("BOOLEAN", {"default": True}),
                 "center_y": ("BOOLEAN", {"default": True}),
-                "video_fps": ("INT",{
-                    "default": 24,
+                "video_fps": ("FLOAT",{
+                    "default": 24.0,
                     "step":1,
                     "display": "number"
                 }),
